feat(yard): move panel into a separate component
